Create SavedStoryButton componenet
